# server/Startup/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chat/create": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "Chat_CreateChatRoom",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "chatRoomDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRoomDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/chat/retreiveChats/patient": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "Chat_GetChatsForPatient",
        "requestBody": {
          "x-name": "chatRoomRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdChatRoomRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/chat/retreiveChats/doctor": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "Chat_GetChatsForDoctor",
        "requestBody": {
          "x-name": "chatRoomRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdChatRoomRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/chat/retreiveMessages": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "Chat_GetChatMessages",
        "requestBody": {
          "x-name": "roomIdRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomIdRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateChatRoomDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "doctorId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserIdChatRoomRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string"
          }
        }
      },
      "RoomIdRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "roomId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}